src/main/java/br/com/caelum/argentum/indicadores/MediaMovelSimples.java
package br.com.caelum.argentum.indicadores;

import br.com.caelum.argentum.modelo.SerieTemporal;

public class MediaMovelSimples implements Indicador {

	private Indicador outroIndicador;

	public double calcula(int posicao, SerieTemporal serie) {

		double soma = 0.0;

		for (int i = posicao - 2; i <= posicao; i++) {
			soma += outroIndicador.calcula(i, serie);
		}

		return soma / 3;

	}

	public MediaMovelSimples(Indicador outroIndicador) {
		this.outroIndicador = outroIndicador;

	}
	
	@Override
	public String toString() {

		return "MMS de " + outroIndicador.toString();

	}

}
